automation/requirements.txt
watchdog>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pdfplumber>=0.10.0
python-docx>=1.1.0
supabase>=2.0.0

